Fix nearest-neighbour bug by giving ids to agents.
When trying to figure out if an agent is happy given a state
we need to find its nearest neighbours in that state. In doing
so we need to exclude the agent itself. The earlier code
had tried to do this by dropping the first element in the list
of neighbours sorted by distance, assuming that that would be
the agent itself at distance 0.

However this does not work if we are checking the happiness of
the agent after it has moved.

We fix this now by giving identities to agents and explicitly
filtering to remove the agent itself from the state before doing the
nearest neighbour calculation.